Revert "Revert "py common.value: Move AbstractValue and Value here" (#13217)"
This reverts commit 7a03e312cea8bfb1ea2177c1ab49312d577d9c7a.
