+++ bindings/pydrake/__init__.py
@@ -77,6 +77,23 @@ def _setattr_kwargs(obj, kwargs):
         setattr(obj, name, value)
 
 
+def _import_cc_module_vars(cc_module, py_module_name):
+    # Imports the cc_module's public symbols into the named py module
+    # (py_module_name), resetting their __module__ to be the py module in the
+    # process.
+    # Returns a list[str] of the public symbol names imported.
+    py_module = sys.modules[py_module_name]
+    var_list = []
+    for name, value in cc_module.__dict__.items():
+        if name.startswith("_"):
+            continue
+        if getattr(value, "__module__", None) == cc_module.__name__:
+            value.__module__ = py_module_name
+        setattr(py_module, name, value)
+        var_list.append(name)
+    return var_list
+
+
 class _DrakeImportWarning(Warning):
     pass
