+++ bindings/pydrake/attic/systems/test/sensors_test.py
@@ -7,13 +7,14 @@ import numpy as np
 
 from pydrake.common import FindResourceOrThrow
 from pydrake.common.eigen_geometry import Isometry3
+from pydrake.common.value import Value
 from pydrake.attic.multibody.rigid_body_tree import (
     AddModelInstancesFromSdfString,
     FloatingBaseType,
     RigidBodyTree,
     RigidBodyFrame,
 )
-from pydrake.systems.framework import InputPort, OutputPort, Value
+from pydrake.systems.framework import InputPort, OutputPort
 
 
 class TestSensors(unittest.TestCase):
