+++ bindings/pydrake/systems/framework_py.cc
@@ -6,7 +6,9 @@
 namespace drake {
 namespace pydrake {
 
-PYBIND11_MODULE(framework, m) {
+// TODO(eric.cousineau): Remove leading underscore once deprecation shim module
+// is removed.
+PYBIND11_MODULE(_framework, m) {
   PYDRAKE_PREVENT_PYTHON3_MODULE_REIMPORT(m);
   m.doc() = "Bindings for the core Systems framework.";
