+++ bindings/pydrake/systems/perception.py
@@ -1,8 +1,9 @@
 import numpy as np
 
+from pydrake.common.value import AbstractValue
 from pydrake.math import RigidTransform
 from pydrake.perception import BaseField, Fields, PointCloud
-from pydrake.systems.framework import AbstractValue, LeafSystem
+from pydrake.systems.framework import LeafSystem
 
 
 def _TransformPoints(points_Ci, X_CiSi):
