+++ bindings/pydrake/systems/test/meshcat_visualizer_test.py
@@ -20,6 +20,7 @@ import meshcat
 import numpy as np
 
 from pydrake.common import FindResourceOrThrow
+from pydrake.common.value import AbstractValue
 from pydrake.geometry import (
     Box,
     GeometryInstance,
@@ -30,7 +31,7 @@ from pydrake.multibody.plant import (
     AddMultibodyPlantSceneGraph)
 from pydrake.multibody.parsing import Parser
 from pydrake.systems.analysis import Simulator
-from pydrake.systems.framework import AbstractValue, DiagramBuilder
+from pydrake.systems.framework import DiagramBuilder
 from pydrake.systems.meshcat_visualizer import (
     MeshcatVisualizer,
     MeshcatContactVisualizer,
