+++ bindings/pydrake/systems/test/sensors_test.py
@@ -6,6 +6,7 @@ import numpy as np
 
 from pydrake.common import FindResourceOrThrow
 from pydrake.common.test_utilities.pickle_compare import assert_pickle
+from pydrake.common.value import AbstractValue, Value
 from pydrake.geometry import FrameId
 from pydrake.geometry.render import CameraProperties, DepthCameraProperties
 from pydrake.math import (
@@ -13,10 +14,8 @@ from pydrake.math import (
     RollPitchYaw,
     )
 from pydrake.systems.framework import (
-    AbstractValue,
     InputPort,
     OutputPort,
-    Value,
     )
 
 # Shorthand aliases, to reduce verbosity.
